sytle : prettier, eslint, airbnb 설정
